% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_predictors.R
\name{contains_NA_left_of_value}
\alias{contains_NA_left_of_value}
\title{Detect if a row has a "__NA__" value left
of a non-"__NA__" value.}
\usage{
contains_NA_left_of_value(...)
}
\arguments{
\item{...}{Values.

 Intended to be used with
 purrr::pmap, where ... is a row of values.}
}
\description{
Detect if a row has a "__NA__" value left
of a non-"__NA__" value.
}
