% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_predictors.R
\name{create_interchangeable_effects_combinations}
\alias{create_interchangeable_effects_combinations}
\title{Creates map of interchangeable effects and their combinations.
Removes all but the first interchangeable effects from fixed_effects.
With this, we can create our formulas with one combination of effects
and add versions with the combinations of interchangeable effects afterwards.}
\usage{
create_interchangeable_effects_combinations(fixed_effects)
}
\arguments{
\item{fixed_effects}{Vector of fixed effects. (Character)}
}
\description{
Creates map of interchangeable effects and their combinations.
Removes all but the first interchangeable effects from fixed_effects.
With this, we can create our formulas with one combination of effects
and add versions with the combinations of interchangeable effects afterwards.
}
