% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{cross_validate}
\alias{cross_validate}
\title{Cross-validate regression models for model selection}
\usage{
cross_validate(model, data, id_column, cat_column = NULL, nfolds = 5,
  family = "gaussian", REML = FALSE, cutoff = 0.5, positive = 1,
  do.plot = FALSE, which_plot = "all", plot_theme = theme_bw(),
  seed = NULL, model_verbose = TRUE)
}

