% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate_list.R
\name{cross_validate_list}
\alias{cross_validate_list}
\title{Cross-validate list of regression models for model selection}
\usage{
cross_validate_list(model_list, data, id_column, cat_column = NULL,
  nfolds = 5, family = "gaussian", REML = FALSE, cutoff = 0.5,
  positive = 1, do.plot = FALSE, seed = NULL, model_verbose = TRUE)
}

