% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_plot.R
\name{cv_plot}
\alias{cv_plot}
\title{Wrapper for plotting common plots using ggplot2}
\usage{
cv_plot(x, type)
}
\arguments{
\item{x}{Object returned by cvms::\link{cross_validate}() (tbl)}

\item{type}{Type of plot.

 \subsection{Gaussian}{

 'RMSE' - boxplot

 'r2' - boxplot

 'IC' - boxplot

 'coefficients' - boxplot
 }

 \subsection{Binomial}{

 "ROC" - ROC curve

 }}
}
\description{
Creates various types of plots based
 the output of cvms::\link{cross_validate}()
}
