% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{package}
\name{cvms-package}
\alias{cvms}
\alias{cvms-package}
\title{cvms: A package for cross-validating regression and classification models}
\description{
Perform (repeated) cross-validation on a list of model formulas. Validate the best model on a validation set.
Perform baseline evaluations on your test set. Generate model formulas by combining your fixed effects.
Evaluate predictions from an external model.
}
\details{
Returns results in a \code{tibble} for easy comparison, reporting and further analysis.

The main functions are:
\code{\link[cvms:cross_validate]{cross_validate()}},
\code{\link[cvms:cross_validate_fn]{cross_validate_fn()}},
\code{\link[cvms:validate]{validate()}},
\code{\link[cvms:validate_fn]{validate_fn()}},
\code{\link[cvms:baseline]{baseline()}},
and \code{\link[cvms:evaluate]{evaluate()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ludvigolsen/cvms}
  \item Report bugs at \url{https://github.com/ludvigolsen/cvms/issues}
}

}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
