% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{package}
\name{cvms}
\alias{cvms}
\alias{cvms-package}
\title{cvms: A package for cross-validating gaussian and binomial regression models}
\description{
Perform repeated cross-validation. Returns results in a tibble for easy comparison, reporting and further analysis
}
\details{
Subsetting methods for balanced cross-validation, time series windowing,
and general grouping and splitting of data.

The cvms package provides 3 main functions:
\code{cross_validate}, \code{validate}, \code{baseline}

And a couple of helper functions:
\code{combine_predictors}, \code{select_metrics}, \code{reconstruct_formulas}, \code{cv_plot}.
}
\section{cross_validate}{

Cross-validate a list of models.

Description of the function

Go to \code{\link{cross_validate}}
}

\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
