% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{evaluate}
\usage{
evaluate(data, dependent_col, prediction_cols, id_col = NULL,
  id_method = "mean", model = NULL, type = "linear_regression",
  apply_softmax = TRUE, cutoff = 0.5, positive = 2,
  metrics = list())
}
\arguments{
\item{type}{Type of evaluation to perform. Allowed values are
\code{"gaussian"} for linear regression;
\code{"binomial"} for binary classification;
\code{"multinomial"} for multiclass classification.}

\item{metrics}{List for enabling/disabling metrics.

  E.g. \code{list("RMSE" = FALSE)} would remove RMSE from the results.
  Default values (TRUE/FALSE) will be used for the remaining metrics available.

  N.B. Currently, disabled metrics are still computed. This is expected to change.}
}
\description{
evaluate
}
\details{
Info to add:
 NB.: precision = positive prediction value
 NB.: recall = sensitivity
 NB.: overall accuracy = micro-F1 = micro-precision = micro-recall
 NB.: Macro F1 is the arithmetic mean of the macro precision and macro recall

 NB.: When type is \code{"multinomial"}, macro-averaging of metrics returns NaN,
 if any of the class level results are NaN.
}
