% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_residuals.R
\name{evaluate_residuals}
\alias{evaluate_residuals}
\title{Evaluate residuals from a regression task}
\usage{
evaluate_residuals(data, predictions_col, targets_col, metrics = list())
}
\arguments{
\item{data}{Data frame with predictions and targets.}

\item{predictions_col}{Name of column with the predicted values in \code{data}.}

\item{targets_col}{Name of the column with the true values in \code{data}.}

\item{metrics}{List for enabling/disabling metrics.

  E.g. \code{list("RMSE" = FALSE)} would disable \code{RMSE}.
  Default values (\code{TRUE}/\code{FALSE}) will be used for the remaining available metrics.

  You can enable/disable all metrics at once by including
  \code{"all" = TRUE/FALSE} in the list. This is done prior to enabling/disabling
  individual metrics, why f.i. \code{list("all" = FALSE, "RMSE" = TRUE)}
  would return only the RMSE metric.

  Also accepts the string \code{"all"}.}
}
\value{
Tbl (tibble) with the calculated metrics.

 The following metrics are available (see \code{metrics}):

 \tabular{rrr}{
  \strong{Metric} \tab \strong{Name} \tab \strong{Default} \cr
  Mean Absolute Error \tab "MAE" \tab Enabled \cr
  Root Mean Square Error \tab "RMSE" \tab Enabled \cr
  Normalized RMSE (by target range) \tab "NRMSE" \tab Disabled \cr
  Normalized RMSE (by target IQR) \tab "RMSEIQR" \tab Disabled \cr
  Normalized RMSE (by target STD) \tab "RMSESTD" \tab Disabled \cr
  Root Mean Squared Log Error \tab "RMSLE" \tab Enabled \cr
  Mean Absolute Log Error \tab "MALE" \tab Disabled \cr
  Relative Absolute Error \tab "RAE" \tab Disabled \cr
  Relative Squared Error \tab "RSE" \tab Disabled \cr
  Root Relative Squared Error \tab "RRSE" \tab Disabled \cr
  Mean Absolute Percentage Error \tab "MAPE" \tab Disabled \cr
  Mean Squared Error \tab "MSE" \tab Disabled \cr
  Total Absolute Error \tab "TAE" \tab Disabled \cr
  Total Squared Error \tab "TSE" \tab Disabled \cr
 }

 The \strong{Name} column refers to the name used in the package.
 This is the name in the output and when enabling/disabling in \code{metrics}.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

 Calculates a large set of error metrics from regression residuals.

 \strong{Note}: Used in \code{\link[cvms:evaluate]{evaluate()}}.
 In most contexts, that is the function you want.
}
\details{
The metric formulas are listed in the \emph{Metrics} vignette.
}
\examples{
# Attach cvms
library(cvms)

# Two classes

# Create targets and predictions
targets <- c(0,1,0,1,0,1,0,1,0,1,0,1)
predictions <- c(1,1,0,0,0,1,1,1,0,1,0,0)

# Create confusion matrix with default metrics
cm <- confusion_matrix(targets, predictions)
cm
cm[["Confusion Matrix"]]
cm[["Table"]]

# Three classes

# Create targets and predictions
targets <- c(0,1,2,1,0,1,2,1,0,1,2,1,0)
predictions <- c(2,1,0,2,0,1,1,2,0,1,2,0,2)

# Create confusion matrix with default metrics
cm <- confusion_matrix(targets, predictions)
cm
cm[["Confusion Matrix"]]
cm[["Table"]]

# Enabling weighted accuracy

# Create confusion matrix with Weighted Accuracy enabled
cm <- confusion_matrix(targets, predictions,
                       metrics = list("Weighted Accuracy" = TRUE))
cm
}
\seealso{
Other evaluation functions: 
\code{\link{confusion_matrix}()},
\code{\link{evaluate}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{evaluation functions}
