% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_model_functions.R
\name{example_model_functions}
\alias{example_model_functions}
\title{Examples of model_fn functions}
\usage{
example_model_functions(name)
}
\arguments{
\item{name}{Name of model to get model function for,
 as it appears in the following list:

 \tabular{rrr}{
  \strong{Name} \tab \strong{Function} \tab \strong{Hyperparameters} \cr
  "lm" \tab \code{\link[stats:lm]{stats::lm()}} \tab \cr
  "lmer" \tab \code{\link[lme4:lmer]{lme4::lmer()}} \tab \code{REML} \cr
  "glm_binomial" \tab \code{\link[stats:lm]{stats::glm()}} \tab \cr
  "glmer_binomial" \tab \code{\link[lme4:glmer]{lme4::glmer()}} \tab \cr
  "svm_gaussian" \tab \code{\link[e1071:svm]{e1071::svm()}} \tab \code{kernel}, \code{cost}\cr
  "svm_binomial" \tab \code{\link[e1071:svm]{e1071::svm()}} \tab \code{kernel}, \code{cost}\cr
  "svm_multinomial" \tab \code{\link[e1071:svm]{e1071::svm()}} \tab \code{kernel}, \code{cost}\cr
  "naive_bayes" \tab \code{\link[e1071:naiveBayes]{e1071::naiveBayes()}} \tab \code{laplace} \cr
  }}
}
\value{
A function with the following form:

 \code{function(train_data, formula, hyperparameters) \{}

 \verb{    }\code{# Return fitted model object}

 \code{\}}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

 Examples of model functions that can be used in
 \code{\link[cvms:cross_validate_fn]{cross_validate_fn()}}.
 They can either be used directly or be starting points.
}
\seealso{
Other example functions: 
\code{\link{example_predict_functions}()},
\code{\link{example_preprocess_functions}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{example functions}
