% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_preprocess_functions.R
\name{example_preprocess_functions}
\alias{example_preprocess_functions}
\title{Examples of preprocess_fn functions}
\usage{
example_preprocess_functions(name)
}
\arguments{
\item{name}{Name of preprocessing function
 as it appears in the following list:

 \tabular{rrr}{
  \strong{Name} \tab \strong{Description} \cr
  "standardize" \tab Centers and scales the numeric variables\cr
  "normalize" \tab Normalizes the numeric variables\cr
  "warn" \tab Identity function that throws a warning and a message\cr
  }}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

 Examples of preprocess functions that can be used in
 \code{\link[cvms:cross_validate_fn]{cross_validate_fn()}}.
 They can either be used directly or be starting points.

 The examples will mainly use
 \code{\link[caret:preProcess]{caret::preProcess()}},
 but you can also use \code{\link[recipes:recipe]{recipes}} or
 similar functions.
}
\seealso{
Other example functions: \code{\link{example_model_functions}},
  \code{\link{example_predict_functions}}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{example functions}
