% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_predictors.R
\name{find_interactions}
\alias{find_interactions}
\title{Finds the interactions in a formula row.
Replaces interactions by value in interactions_map.
Returns data frame row with effects and interactions as columns
and "1" as value.}
\usage{
find_interactions(..., interactions_map)
}
\arguments{
\item{...}{Values.

 Intended to be used with
 purrr::pmap, where ... is a row of values.}

\item{interactions_map}{Named list with each
potential interaction mapping to its key permutation.}
}
\description{
Finds the interactions in a formula row.
Replaces interactions by value in interactions_map.
Returns data frame row with effects and interactions as columns
and "1" as value.
}
