% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{font}
\alias{font}
\title{Create a list of font settings for plots}
\usage{
font(
  size = NULL,
  color = NULL,
  alpha = NULL,
  nudge_x = NULL,
  nudge_y = NULL,
  angle = NULL,
  family = NULL,
  fontface = NULL,
  hjust = NULL,
  vjust = NULL,
  lineheight = NULL,
  digits = NULL,
  prefix = NULL,
  suffix = NULL
)
}
\arguments{
\item{size, color, alpha, nudge_x, nudge_y, angle, family, fontface, hjust, vjust, lineheight}{Either values directly passed to
\code{\link[ggplot2:geom_text]{ggplot2::geom_text}} or functions that take in the values (e.g., counts, percentages, etc.)
and returns a vector of values to pass to \code{\link[ggplot2:geom_text]{ggplot2::geom_text}} via
\code{aes()}.}

\item{digits}{Number of digits to round to. If negative, no rounding will take place.}

\item{prefix}{A string prefix.}

\item{suffix}{A string suffix.}
}
\value{
List of settings.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

 Creates a list of font settings for plotting with cvms plotting functions.

 Some arguments can take either the value to use directly
 OR a function that takes one argument (vector with the
 values to set a font for; e.g., the counts, percentages, etc.)
 and returns the value(s) to use for each element. Such a
 function could for instance specify different font
 colors for different background intensities.

 NOTE: This is experimental and could change.
}
\seealso{
Other plotting functions: 
\code{\link{dynamic_font_color_settings}()},
\code{\link{plot_confusion_matrix}()},
\code{\link{plot_metric_density}()},
\code{\link{plot_probabilities}()},
\code{\link{plot_probabilities_ecdf}()},
\code{\link{sum_tile_settings}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{plotting functions}
