% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosing_metrics_functions.R
\name{gaussian_metrics}
\alias{gaussian_metrics}
\title{Select metrics for Gaussian evaluation}
\usage{
gaussian_metrics(
  all = NULL,
  rmse = NULL,
  mae = NULL,
  nrmse = NULL,
  rmseiqr = NULL,
  rmsestd = NULL,
  rmsle = NULL,
  male = NULL,
  rae = NULL,
  rse = NULL,
  rrse = NULL,
  mape = NULL,
  mse = NULL,
  tae = NULL,
  tse = NULL,
  r2m = NULL,
  r2c = NULL,
  aic = NULL,
  aicc = NULL,
  bic = NULL
)
}
\arguments{
\item{all}{Enable/disable all arguments at once. (Logical)

 Specifying other metrics will overwrite this, why you can
 use (\code{all = FALSE, rmse = TRUE}) to get only the \code{RMSE} metric.}

\item{rmse}{\code{RMSE}. (Default: TRUE)}

\item{mae}{\code{MAE}. (Default: TRUE)}

\item{nrmse}{\code{NRMSE}. (Default: FALSE)}

\item{rmseiqr}{\code{RMSEIQR}. (Default: FALSE)}

\item{rmsestd}{\code{RMSESTD}. (Default: FALSE)}

\item{rmsle}{\code{RMSLE}. (Default: TRUE)}

\item{male}{\code{MALE}. (Default: FALSE)}

\item{rae}{\code{RAE}. (Default: FALSE)}

\item{rse}{\code{RSE}. (Default: FALSE)}

\item{rrse}{\code{RRSE}. (Default: FALSE)}

\item{mape}{\code{MAPE}. (Default: FALSE)}

\item{mse}{\code{MSE}. (Default: FALSE)}

\item{tae}{\code{TAE}. (Default: FALSE)}

\item{tse}{\code{TSE}. (Default: FALSE)}

\item{r2m}{\code{r2m}. (Default: FALSE)}

\item{r2c}{\code{r2c}. (Default: FALSE)}

\item{aic}{\code{AIC}. (Default: FALSE)}

\item{aicc}{\code{AICc}. (Default: FALSE)}

\item{bic}{\code{BIC}. (Default: FALSE)}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

 Enable/disable metrics for Gaussian evaluation. Can be supplied to the
 \code{metrics} argument in many of the \code{cvms} functions.

 Note: Some functions may have slightly different defaults than the ones supplied here.
}
\examples{
\donttest{
# Attach packages
library(cvms)

# Enable only RMSE
gaussian_metrics(all = FALSE, rmse = TRUE)

# Enable all but RMSE
gaussian_metrics(all = TRUE, rmse = FALSE)

# Disable RMSE
gaussian_metrics(rmse = FALSE)
}
}
\seealso{
Other evaluation functions: 
\code{\link{binomial_metrics}()},
\code{\link{confusion_matrix}()},
\code{\link{evaluate_residuals}()},
\code{\link{evaluate}()},
\code{\link{multinomial_metrics}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{evaluation functions}
