% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_predictors.R
\name{generate_interactions}
\alias{generate_interactions}
\title{Generate interactions (One permutation only)
Add columns describing if an effect is in the interaction}
\usage{
generate_interactions(possible_interactions)
}
\arguments{
\item{possible_interactions}{Data Frame.}
}
\description{
Generate interactions (One permutation only)
Add columns describing if an effect is in the interaction
}
