% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_predictors.R
\name{generate_interactions_map}
\alias{generate_interactions_map}
\title{Generate interactions map with all permutations of
an interaction being mapped to one of the permutations}
\usage{
generate_interactions_map(possible_interactions)
}
\arguments{
\item{possible_interactions}{Data Frame.}
}
\description{
Generate interactions map with all permutations of
an interaction being mapped to one of the permutations
}
