% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_predictors.R
\name{get_max_nway_interaction}
\alias{get_max_nway_interaction}
\title{Finds the largest interaction.}
\usage{
get_max_nway_interaction(...)
}
\arguments{
\item{...}{Values.

 Intended to be used with
 purrr::pmap, where ... is a row of values.}
}
\description{
Finds the largest interaction.
}
