% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{musicians}
\alias{musicians}
\title{Musician groups}
\format{
A data frame with 60 rows and 9 variables:
\describe{
  \item{ID}{Musician identifier, 60 levels}
  \item{Age}{Age of the musician. Between 17 and 66 years.}
  \item{Class}{The class of the musician. One of \code{"A"}, \code{"B"}, \code{"C"}, and \code{"D"}.}
  \item{Height}{Height of the musician. Between \code{146} and \code{196} centimeters.}
  \item{Drums}{Whether the musician plays drums. \code{0} = No, \code{1} = Yes.}
  \item{Bass}{Whether the musician plays bass. \code{0} = No, \code{1} = Yes.}
  \item{Guitar}{Whether the musician plays guitar. \code{0} = No, \code{1} = Yes.}
  \item{Keys}{Whether the musician plays keys. \code{0} = No, \code{1} = Yes.}
  \item{Vocals}{Whether the musician sings. \code{0} = No, \code{1} = Yes.}
}
}
\description{
Made-up data on 60 musicians in 4 groups for multiclass classification.
}
\seealso{
predicted.musicians
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\keyword{data}
