% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{participant.scores}
\alias{participant.scores}
\title{Participant scores}
\description{
10 participants. 5 for each of 2 diagnoses.
Scores for 3 sessions per participant. Participants get better for each session.
Age of participants.
}
\details{
Simulated data. 30 rows. 5 columns.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\keyword{data}
