% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_predictors.R
\name{paste_as_interactions}
\alias{paste_as_interactions}
\title{Paste effects as interactions.
Collapse with ".|_|." instead of " * ".}
\usage{
paste_as_interactions(...)
}
\arguments{
\item{...}{Values.

 Intended to be used with
 purrr::pmap, where ... is a row of values.}
}
\description{
Paste effects as interactions.
Collapse with ".|_|." instead of " * ".
}
