% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_metric_density}
\alias{plot_metric_density}
\title{Density plot for a metric}
\usage{
plot_metric_density(
  data,
  metric = "Overall Accuracy",
  fill = "pink",
  alpha = 0.4,
  theme_fn = ggplot2::theme_light,
  xlim = NULL
)
}
\arguments{
\item{metric}{Name of the metric column in \code{data} to plot. (Character)}

\item{fill}{Color of the plotted distribution.}

\item{alpha}{Transparency of the distribution (0 - 1).}

\item{theme_fn}{The ggplot2 theme function to apply.}

\item{xlim}{Limits for the x-axis. Can be set to NULL.

 E.g. \code{c(0, 1)}.}

\item{Data}{frame with a metric column to create density plot for.}
}
\value{
ggplot2 object.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

 Creates a ggplot2 object with a density plot for one of the columns in the passed data frame.
}
\seealso{
Other plotting functions: 
\code{\link{font}()},
\code{\link{plot_confusion_matrix}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{plotting functions}
