% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{precomputed.formulas}
\alias{precomputed.formulas}
\title{Precomputed formulas}
\format{
A data frame with 259,358 rows and 5 variables:
\describe{
  \item{formula_}{combination of fixed effects, separated by "\code{+}" and "\code{*}"}
  \item{max_interaction_size}{maximum interaction size in the formula, up to \code{3}}
  \item{max_effect_frequency}{maximum count of an effect in the formula, e.g. the \code{3} A's in \code{"A * B + A * C + A * D"}}
  \item{num_effects}{number of unique effects included in the formula}
  \item{min_num_fixed_effects}{minimum number of fixed effects required to use the formula,
  i.e. the index in the alphabet of the last of the alphabetically ordered effects (letters) in the formula,
  so \code{4} for the formula: \code{"A + B + D"}  }
}
}
\description{
Fixed effect combinations for model formulas with/without two- and three-way interactions.
Up to eight fixed effects in total with up to five fixed effects per formula.
}
\details{
Effects are represented by the first eight capital letters.

Used by \code{\link[cvms:combine_predictors]{combine_predictors}}.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\keyword{data}
