% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_predictors.R
\name{remove_duplicates_and_rows_without_interactions}
\alias{remove_duplicates_and_rows_without_interactions}
\title{Deduplicates rows with the same interactions (including permutations).
Removes rows without interactions, as we got these covered already.}
\usage{
remove_duplicates_and_rows_without_interactions(formulas_df,
  interactions_map)
}
\arguments{
\item{formulas_df}{Data Frame where each column is an effect or an operator}

\item{interactions_map}{Named list with each
potential interaction mapping to its key permutation.}
}
\description{
Deduplicates rows with the same interactions (including permutations).
Removes rows without interactions, as we got these covered already.
}
