% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_metrics.R
\name{summarize_metrics}
\alias{summarize_metrics}
\title{Summarize metrics with common descriptors}
\usage{
summarize_metrics(data, cols = NULL, na.rm = TRUE, inf.rm = TRUE)
}
\arguments{
\item{data}{Data frame with numeric columns to summarize.}

\item{cols}{Names of columns to summarize. Non-numeric columns are ignored.}

\item{na.rm}{Whether to remove NAs before summarizing. (logical)}

\item{inf.rm}{Whether to remove Infs before summarizing. (logical)}
}
\value{
tibble where each row is a descriptor of the column.

 The \strong{Measure} column contains the name of the descriptor.

 The \strong{NAs} row is a count of \code{NA}s in the column.

 The \strong{INFs} row is a count of \code{Inf}s in the column.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

 Summarizes all numeric columns. Counts the NAs and Infs in the columns.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
