% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate regression model on test set}
\usage{
validate(train_data, models, test_data = NULL,
  partitions_col = ".partitions", family = "gaussian", link = NULL,
  control = NULL, REML = FALSE, cutoff = 0.5, positive = 1,
  err_nc = FALSE, rm_nc = FALSE, model_verbose = FALSE)
}
\arguments{
\item{train_data}{Dataframe.}

\item{models}{Model formulas as strings. (Character)

 E.g. \code{c("y~x", "y~z")}.

 Can contain random effects.

 E.g. \code{c("y~x+(1|r)", "y~z+(1|r)")}.}

\item{test_data}{Dataframe. If specifying \code{partitions_col}, this can be \code{NULL}.}

\item{partitions_col}{Name of grouping factor for identifying partitions. (Character)
 1 is training set and 2 is test set.

 Only used if test_data is NULL.}

\item{family}{Name of family ('gaussian' or 'binomial'). (Character)}

\item{link}{Link function. (Character)

 E.g. \code{link = "log"} with \code{family = "gaussian"} will
 use \code{family = gaussian(link = "log")}.

 See \link[stats]{family} for available link functions.

 \subsection{Default link functions}{

 Gaussian: \code{'identity'}.

 Binomial: \code{'logit'}.}}

\item{control}{Construct control structures for mixed model fitting (i.e. \link[lme4]{lmer} and \link[lme4]{glmer}).
See \link[lme4]{lmerControl} and \link[lme4]{glmerControl}.
Ignored if fitting \code{lm} and \code{glm} models.}

\item{REML}{Restricted Maximum Likelihood. (Logical)}

\item{cutoff}{Threshold for predicted classes. Binomial models only. (Numeric)}

\item{positive}{Level from dependent variable to predict (1 or 2 - alphabetically).
Used when creating confusion matrix.
Binomial models only. (Integer)}

\item{err_nc}{Raise error if model does not converge. (Logical)}

\item{rm_nc}{Remove non-converged models from output. (Logical)}

\item{model_verbose}{Print name of used model function on each iteration. (Logical)}
}
\value{
List containing tbl (tibble) with results and the trained model object.
 The tibble contains:

 \subsection{Gaussian Results}{
 \strong{RMSE}, \strong{r2m}, \strong{r2c}, \strong{AIC}, \strong{AICc},
 and \strong{BIC}.

 Count of \strong{convergence warnings}. Consider discarding the model if it did not converge.

 Specified \strong{family}.

 A tibble with \strong{coefficients} of the model.

 Name of \strong{dependent} variable.

 Names of \strong{fixed} effects.

 Names of \strong{random} effects if any.

 }

 \subsection{Binomial Results}{
 Based on predictions of the test set,
 a confusion matrix and ROC curve are used to get the following:

 ROC:

 \strong{AUC}, \strong{Lower CI}, and \strong{Upper CI}

 Confusion Matrix:

 \strong{Kappa}, \strong{Sensitivity},
 \strong{Specificity}, \strong{Positive Prediction Value},
 \strong{Negative Prediction Value},
 \strong{F1}, \strong{Prevalence}, \strong{Detection Rate},
 \strong{Detection Prevalence}, and
 \strong{Balanced Accuracy}.

 Count of \strong{convergence warnings}. Consider discarding the model if it did not converge.

 Specified \strong{family}.

 A tibble with \strong{predictions}, predicted classes (depends on \code{cutoff}), and the targets.

 A tibble with the sensativities and specificities from the \strong{ROC} curve.

 Name of \strong{dependent} variable.

 Names of \strong{fixed} effects.

 Names of \strong{random} effects if any.

 }
}
\description{
Train gaussian or binomial models on a full training set and validate it by predicting the test/validation set.
 Returns results in a tibble for easy reporting, along with the trained models.

 \strong{validate() is under development! Large changes may occur.}
}
\details{
\subsection{Models}{

 Gaussian: stats::lm, lme4::lmer

 Binomial: stats::glm, lme4::glmer
 }
 \subsection{Results}{
 \strong{Gaussian}:

 RMSE : hydroGOF::rmse

 r2m : MuMIn::r.squaredGLMM

 r2c : MuMIn::r.squaredGLMM

 AIC : stats::AIC

 AICc : AICcmodavg::AICc

 BIC : stats::BIC

 \strong{Binomial}:

 Confusion matrix: caret::confusionMatrix

 ROC: pROC::roc
 }
}
\examples{
# Attach libraries
library(cvms)
library(groupdata2) # partition()
library(dplyr) # \%>\% arrange()

# Data is part of cvms
data <- participant.scores

# Set seed for reproducibility
set.seed(7)

# Partition data
# Keep as single dataframe
# We could also have fed validate() separat train and test sets.
data_partitioned <- partition(data,
                              p = 0.7,
                              cat_col = 'diagnosis',
                              id_col = 'participant',
                              list_out=FALSE) \%>\%
    arrange(.partitions)

# Validate a model

# Gaussian
validate(data_partitioned,
         models = "score~diagnosis",
         partitions_col = '.partitions',
         family='gaussian',
         REML = FALSE)

# Binomial
validate(data_partitioned,
         models = "diagnosis~score",
         partitions_col = '.partitions',
         family='binomial')

# Use non-default link functions

validate(data_partitioned,
         models = "score~diagnosis",
         partitions_col = '.partitions',
         family = 'gaussian',
         link = 'log',
         REML = FALSE)

## Feed separat train and test sets

# Partition data to list of dataframes
# The first dataframe will be train (70\% of the data)
# The second will be test (30\% of the data)
data_partitioned <- partition(data, p = 0.7,
                              cat_col = 'diagnosis',
                              id_col = 'participant',
                              list_out=TRUE)
train_data <- data_partitioned[[1]]
test_data <- data_partitioned[[2]]

# Validate a model

# Gaussian
validate(train_data,
         test_data = test_data,
         models = "score~diagnosis",
         family='gaussian',
         REML = FALSE)

}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}

Benjamin Hugh Zachariae
}
