% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_predictors.R
\name{vals_to_cols}
\alias{vals_to_cols}
\title{Create data frame row with values as columns.
Fill with the values or a specified value.}
\usage{
vals_to_cols(..., fill_with = NULL)
}
\arguments{
\item{...}{Values.

 Intended to be used with
 purrr::pmap, where ... is a row of values.}

\item{fill_with}{Value to fill cells with.
If \code{NULL}, the values in \code{...} are used.}
}
\description{
Create data frame row with values as columns.
Fill with the values or a specified value.
}
