% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{wines}
\alias{wines}
\title{Wine varieties}
\format{A data frame with 372 rows and 1 variable:
\describe{
  \item{Variety}{Wine variety, 11 levels}
}}
\description{
A list of wine varieties in an approximately Zipfian distribution.
}
\details{
Based on the wine-reviews (v4) kaggle dataset by Zack Thoutt:
https://www.kaggle.com/zynicide/wine-reviews
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\keyword{data}
