% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_predictors.R
\name{combine_predictors}
\alias{combine_predictors}
\title{Create model formulas with all combinations of the fixed effects
Random effects will be added as supplied}
\usage{
combine_predictors(dependent, fixed_effects, random_effects = NULL,
  max_fixed_effects = NULL, max_interaction_size = NULL)
}
\description{
Create model formulas with all combinations of the fixed effects
Random effects will be added as supplied
}
