% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_predictors.R
\name{combine_predictors}
\alias{combine_predictors}
\alias{generate_formulas}
\title{Generate model formulas by combining predictors}
\usage{
combine_predictors(dependent, fixed_effects, random_effects = NULL,
  max_fixed_effects = NULL, max_interaction_size = NULL)
}
\arguments{
\item{dependent}{Name of dependent variable. (Character)}

\item{fixed_effects}{List of fixed effects. (Character)

 Effects in sublists will be interchanged. This can be useful, when
 we have multiple versions of a predictor (e.g. \code{x1} and \code{log(x1)}) that we
 do not wish to have in the same formula.

 Example of interchangeable effects:

 \code{list( list( "x1", "log_x1" ), "x2", "x3" )}}

\item{random_effects}{The random effects structure. (Character)

 Is appended to the model formulas.}

\item{max_fixed_effects}{Maximum number of fixed effects in a model formula. (Integer)

  Due to the exponential time increase with number of fixed effects,
  it can help to restrict the number of fixed effects in a formula.}

\item{max_interaction_size}{Maximum number of effects in an interaction. (Integer)

 Use this to limit the \code{n}-way interactions allowed.

 A model formula can contain multiple interactions.}
}
\value{
List of model formulas.

 E.g.:

 \code{c("y ~ x1 + (1|z)", "y ~ x2 + (1|z)",
 "y ~ x1 + x2 + (1|z)", "y ~ x1 * x2 + (1|z)")}.
}
\description{
Create model formulas with every combination
 of your fixed effects, along with the dependent variable and random effects.
 Be aware of exponential time increase with number of fixed effects.
 Allows for limiting the size of interactions and the number of fixed effects in a formula.
 Currently use the \code{+} and \code{*} operators,
 and has the constraint that a fixed effect is only included once in a formula.
}
\examples{
# Attach libraries
library(cvms)

# Create effect names
dependent <- "y"
fixed_effects <- c("a","b","c","d")
random_effects <- "(1|e)"

# Create model formulas
combine_predictors(dependent, fixed_effects,
                   random_effects)

# Create effect names with interchangeable effects in sublists
fixed_effects <- list("a",list("b","log_b"),"c","d")

# Create model formulas
combine_predictors(dependent, fixed_effects,
                   random_effects)
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
