% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{package}
\name{cvms}
\alias{cvms}
\alias{cvms-package}
\title{cvms: A package for cross-validating gaussian and binomial regression models}
\description{
Perform (repeated) cross-validation on a list of model formulas. Validate the best model on a validation set.
Perform baseline evaluations on your test set. Generate model formulas by combining your fixed effects.
}
\details{
Returns results in a tibble for easy comparison, reporting and further analysis.

The cvms package provides 3 main functions:
\code{cross_validate}, \code{validate}, \code{baseline}.

And a couple of helper functions:
\code{combine_predictors}, \code{select_metrics}, \code{reconstruct_formulas}, \code{cv_plot}.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
