% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{font}
\alias{font}
\title{Create a list of font settings for plots}
\usage{
font(
  size = NULL,
  color = NULL,
  alpha = NULL,
  nudge_x = NULL,
  nudge_y = NULL,
  angle = NULL,
  family = NULL,
  fontface = NULL,
  hjust = NULL,
  vjust = NULL,
  lineheight = NULL,
  digits = NULL,
  prefix = NULL,
  suffix = NULL
)
}
\arguments{
\item{size, color, alpha, nudge_x, nudge_y, angle, family, fontface, hjust, vjust, lineheight}{As passed to
\code{\link[ggplot2:geom_text]{ggplot2::geom_text}}.}

\item{digits}{Number of digits to round to. If negative, no rounding will take place.}

\item{prefix}{A string prefix.}

\item{suffix}{A string suffix.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

 Creates a list of font settings for plotting with cvms plotting functions.

 NOTE: This is very experimental and will likely change.
}
\seealso{
Other plotting functions: 
\code{\link{plot_confusion_matrix}()},
\code{\link{plot_metric_density}()},
\code{\link{plot_probabilities}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{plotting functions}
