% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{font}
\alias{font}
\title{Create a list of font settings for plots}
\usage{
font(
  size = NULL,
  color = NULL,
  nudge_x = NULL,
  nudge_y = NULL,
  angle = NULL,
  digits = NULL,
  prefix = NULL,
  suffix = NULL
)
}
\arguments{
\item{size}{Font size.}

\item{color}{Color of the text. As passed to
\code{\link[ggplot2:geom_text]{ggplot2::geom_text}}.}

\item{nudge_x, nudge_y}{Nudge on x/y axes. As passed to
\code{\link[ggplot2:geom_text]{ggplot2::geom_text}}.}

\item{angle}{Rotation angle. As passed to
\code{\link[ggplot2:geom_text]{ggplot2::geom_text}}.}

\item{digits}{Number of digits to round to. If negative, no rounding will take place.}

\item{prefix}{A string prefix.}

\item{suffix}{A string suffix.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

 Creates a list of font settings for plotting with cvms plotting functions.

 NOTE: This is very experimental and will likely change.
}
\seealso{
Other plotting functions: 
\code{\link{plot_confusion_matrix}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{plotting functions}
