% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_challenging.R
\name{most_challenging}
\alias{most_challenging}
\alias{hardest}
\title{Find the data points that were hardest to predict}
\usage{
most_challenging(
  data,
  type,
  obs_id_col = "Observation",
  target_col = "Target",
  prediction_col = ifelse(type == "gaussian", "Prediction", "Predicted Class"),
  threshold = 0.15,
  threshold_is = "percentage"
)
}
\arguments{
\item{data}{Data frame with predictions, targets and observation IDs.

 Can be grouped by \code{\link[dplyr:group_by]{dplyr::group_by()}}.}

\item{type}{Type of task used to get the predictions:

 \code{"gaussian"} for regression (like linear regression).

 \code{"binomial"} for binary classification.

 \code{"multinomial"} for multiclass classification.}

\item{obs_id_col}{Name of column with observation IDs. This will be used to aggregate
the performance of each observation.}

\item{target_col}{Name of column with the true classes/values in \code{data}.}

\item{prediction_col}{Name of column with the predictions.

 When evaluating a classification task, the column should contain the predicted classes.}

\item{threshold}{Threshold to filter observations by. Depends on \code{type} and \code{threshold_is}.

 The \code{threshold} can either be a \strong{percentage} or a \strong{score}.
 For percentages, a lower \code{threshold}
 returns fewer observations. For scores, this depends on \code{type}.

 \subsection{"gaussian"}{
 \subsection{threshold_is "percentage"}{
 (Approximate) percentage of the observations with the largest root mean square errors
 to return.
 }
 \subsection{threshold_is "score"}{
 Observations with a root mean square error larger than or equal to the \code{threshold} will be returned.
 }
 }
 \subsection{"binomial", "multinomial"}{
 \subsection{threshold_is "percentage"}{
 (Approximate) percentage of the observations with the lowest accuracies to return.
 }
 \subsection{threshold_is "score"}{
 Observations with an accuracy below or equal to the threshold will be returned.
 }
 }}

\item{threshold_is}{Either \code{"score"} or \code{"percentage"}. See \code{threshold}.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
 Finds the data points that, overall, were the most challenging to predict,
 based on a prediction score (\code{RMSE} for regression and
 \code{Accuracy} for classification).
}
\examples{
# Attach packages
library(cvms)
library(dplyr)

\donttest{

## Multinomial

# Find the most challenging data points (per classifier)
# in the predicted.musicians dataset
# which resembles the "Predictions" tibble from the evaluation results

# The 20\% lowest scoring on accuracy
most_challenging(predicted.musicians,
                 obs_id_col = "ID",
                 type = "multinomial",
                 threshold = 0.30,
                 threshold_is = "percentage")

# The 20\% lowest scoring on accuracy per classifier
predicted.musicians \%>\%
  dplyr::group_by(Classifier) \%>\%
  most_challenging(obs_id_col = "ID",
                   type = "multinomial",
                   threshold = 0.40,
                   threshold_is = "percentage")

# Accuracy scores below 0.05
most_challenging(predicted.musicians,
                 obs_id_col = "ID",
                 type = "multinomial",
                 threshold = 0.05,
                 threshold_is = "score")

# Accuracy scores below 0.05 per classifier
predicted.musicians \%>\%
  dplyr::group_by(Classifier) \%>\%
  most_challenging(obs_id_col = "ID",
                   type = "multinomial",
                   threshold = 0.05,
                   threshold_is = "score")

## Gaussian

set.seed(1)

df <- data.frame("Observation" = rep(1:10, n = 3),
                 "Target" = rnorm(n = 30, mean = 25, sd = 5),
                 "Prediction" = rnorm(n = 30, mean = 27, sd = 7))

# The 20\% highest RMSE scores
most_challenging(df,
                 type = "gaussian",
                 threshold = 0.2,
                 threshold_is = "percentage")

# RMSE scores above 9
most_challenging(df,
                 type = "gaussian",
                 threshold = 9,
                 threshold_is = "score")
}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
