% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_challenging.R
\name{most_challenging}
\alias{most_challenging}
\alias{hardest}
\title{Find the data points that were hardest to predict}
\usage{
most_challenging(
  data,
  obs_id_col = "Observation",
  target_col = "Target",
  prediction_col = dplyr::case_when(type == "gaussian" ~ "Prediction", TRUE ~
    "Predicted Class"),
  type = "binomial",
  threshold = 0.15,
  threshold_is = "percentage"
)
}
\arguments{
\item{data}{Data frame with predictions, targets and observation IDs.

 Can be grouped by \code{\link[dplyr:group_by]{dplyr::group_by()}}.}

\item{obs_id_col}{Name of column with observation IDs. This will be used to aggregate
the performance of each observation.}

\item{target_col}{Name of column with the true classes/values in \code{data}.}

\item{prediction_col}{Name of column with the predictions.

 When evaluating a classification task, the column should contain the predicted classes.}

\item{type}{Type of task used to get the predictions:

 \code{"gaussian"} for regression (like linear regression).

 \code{"binomial"} for binary classification.

 \code{"multinomial"} for multiclass classification.}

\item{threshold}{Threshold to filter observations by. Depends on \code{type} and \code{threshold_is}.

 The \code{threshold} can either be a \strong{percentage} or a \strong{score}.
 For percentages, a lower \code{threshold}
 returns fewer observations. For scores, this depends on \code{type}.

 \subsection{"gaussian"}{
 \subsection{threshold_is "percentage"}{
 Percentage of the observations with the largest root mean square errors
 to return.
 }
 \subsection{threshold_is "score"}{
 Observations with a root mean square error larger than the \code{threshold} will be returned.
 }
 }
 \subsection{"binomial", "multinomial"}{
 \subsection{threshold_is "percentage"}{
 Percentage of the observations with the lowest accuracies to return.
 }
 \subsection{threshold_is "score"}{
 Observations with an accuracy below the threshold will be returned.
 }
 }}

\item{threshold_is}{Either \code{"score"} or \code{"percentage"}. See \code{threshold}.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
 Finds the data points that, overall, were the most challenging to predict,
 based on a prediction score (\code{RMSE} for regression and
 \code{Accuracy} for classification).
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
