% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confusion_matrix.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot a confusion matrix}
\usage{
plot_confusion_matrix(
  conf_matrix,
  targets_col = "Target",
  predictions_col = "Prediction",
  counts_col = "N",
  add_counts = TRUE,
  add_normalized = TRUE,
  add_horizontal_percentage = TRUE,
  add_vertical_percentage = TRUE,
  counts_on_top = FALSE,
  palette = "Greens",
  theme_fn = ggplot2::theme_light,
  place_x_axis_above = TRUE,
  rotate_y_text = TRUE,
  font_counts = font(),
  font_normalized = font(),
  font_horizontal = font(),
  font_vertical = font(),
  darkness = 0.8
)
}
\arguments{
\item{conf_matrix}{Confusion matrix tibble with each combination of targets and predictions along with their counts.

 E.g. for a binary classification:

 \tabular{rrr}{
  \strong{Target} \tab \strong{Prediction} \tab \strong{N} \cr
  class_1 \tab class_1 \tab 5 \cr
  class_1 \tab class_2 \tab 9 \cr
  class_2 \tab class_1 \tab 3 \cr
  class_2 \tab class_2 \tab 2 \cr
 }

 As created with the various evaluation functions in \code{cvms}, like
 \code{\link[cvms:evaluate]{evaluate()}}.}

\item{targets_col}{Name of column with target levels.}

\item{predictions_col}{Name of column with prediction levels.}

\item{counts_col}{Name of column with a count for each combination
of the target and prediction levels.}

\item{add_counts}{Add the counts to the middle of the tiles. (Logical)}

\item{add_normalized}{Normalize the counts to percentages and
add to the middle of the tiles. (Logical)}

\item{add_horizontal_percentage}{Add the horizontal percentages,
i.e. how big a part of its row the tile makes up. (Logical)}

\item{add_vertical_percentage}{Add the vertical percentages,
i.e. how big a part of its column the tile makes up. (Logical)}

\item{counts_on_top}{Switch the counts and normalized counts,
such that the counts are on top. (Logical)}

\item{palette}{Color scheme. Passed directly to \code{palette} in
\code{\link[ggplot2:scale_fill_distiller]{ggplot2::scale_fill_distiller}}.}

\item{theme_fn}{The ggplot2 theme function to apply.}

\item{place_x_axis_above}{Move the x-axis text to the top and reverse the levels such that}

\item{rotate_y_text}{Whether to rotate the y-axis text to
be vertical instead of horizontal. (Logical)}

\item{font_counts}{List of font settings for the counts.
Can be provided with \code{\link[cvms:font]{font()}}.}

\item{font_normalized}{List of font settings for the normalized counts.
Can be provided with \code{\link[cvms:font]{font()}}.}

\item{font_horizontal}{List of font settings for the horizontal percentages.
Can be provided with \code{\link[cvms:font]{font()}}.}

\item{font_vertical}{List of font settings for the vertical percentages.
Can be provided with \code{\link[cvms:font]{font()}}.}

\item{darkness}{How dark the darkest colors should be, between 0 and 1, where 1 is darkest.

 Technically, a lower value increases the upper limit in
 \code{\link[ggplot2:scale_fill_distiller]{ggplot2::scale_fill_distiller}}.}
}
\value{
A ggplot2 object representing a confusion matrix. Color intensity depends on the counts.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

 Creates a ggplot2 object representing a confusion matrix with counts,
 percentages overall, horizontal percentages and vertical percentages.
}
\details{
Inspired by Antoine Sachet at https://stackoverflow.com/a/53612391/11832955
}
\examples{
# Attach cvms
library(cvms)

# Two classes

# Create targets and predictions
targets <- c(0,1,0,1,0,1,0,1,0,1,0,1)
predictions <- c(1,1,0,0,0,1,1,1,0,1,0,0)

# Create confusion matrix with default metrics
cm <- confusion_matrix(targets, predictions)
cm[["Confusion Matrix"]]

plot_confusion_matrix(cm[["Confusion Matrix"]][[1]])

# Three (or more) classes

# Create targets and predictions
targets <- c(0,1,2,1,0,1,2,1,0,1,2,1,0)
predictions <- c(2,1,0,2,0,1,1,2,0,1,2,0,2)

# Create confusion matrix with default metrics
cm <- confusion_matrix(targets, predictions)
cm[["Confusion Matrix"]]

plot_confusion_matrix(cm[["Confusion Matrix"]][[1]])
}
\seealso{
Other plotting functions: 
\code{\link{font}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{plotting functions}
