% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_metric_density}
\alias{plot_metric_density}
\title{Density plot for a metric}
\usage{
plot_metric_density(
  results = NULL,
  baseline = NULL,
  metric = "",
  fill = c("darkblue", "lightblue"),
  alpha = 0.6,
  theme_fn = ggplot2::theme_minimal,
  xlim = NULL
)
}
\arguments{
\item{results}{Data frame with a metric column to create density plot for.

 To only plot the baseline, set to \code{NULL}.}

\item{baseline}{Data frame with the random evaluations from \code{\link[cvms:baseline]{baseline()}}.
 Should contain a column for the \code{metric}.

 To only plot the results, set to \code{NULL}.}

\item{metric}{Name of the metric column in \code{results} to plot. (Character)}

\item{fill}{Colors of the plotted distributions.
The first color is for the \code{baseline}, the second for the \code{results}.}

\item{alpha}{Transparency of the distribution (0 - 1).}

\item{theme_fn}{The ggplot2 theme function to apply.}

\item{xlim}{Limits for the x-axis. Can be set to NULL.

 E.g. \code{c(0, 1)}.}
}
\value{
ggplot2 object.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

 Creates a ggplot2 object with a density plot for one of the columns in the passed data frame.
}
\seealso{
Other plotting functions: 
\code{\link{font}()},
\code{\link{plot_confusion_matrix}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{plotting functions}
