% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_formulas.R
\name{reconstruct_formulas}
\alias{reconstruct_formulas}
\title{Reconstruct model formulas from results tibbles}
\usage{
reconstruct_formulas(results, topn = NULL)
}
\arguments{
\item{results}{Data frame with results from cross_validate or validate(). (tbl)

 Must contain at least the columns "Dependent" and "Fixed". For random effects
 the "Random" column should be included.}

\item{topn}{Number of top rows to return. Simply applies head() to the results tibble.

 Set to NULL to return all rows.}
}
\value{
List of model formulas.
}
\description{
In the results tibble from cross_validate and validate,
 the model formulas have been split into the columns Dependent, Fixed and Random.
 Quickly reconstruct the model formulas from these column.

 NOTE: Currently only works properly without random effects!
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
