% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_metrics.R
\name{select_metrics}
\alias{select_metrics}
\title{Select columns with evaluation metrics and model definitions.}
\usage{
select_metrics(results)
}
\arguments{
\item{results}{Results tibble from cross_validate() or validate().}
}
\value{
The results tibble with only metric and model definition columns.
}
\description{
When reporting results, we may not want all
 the nested tibbles and process information columns.
 This function selects the evaluation metrics and model formulas only.
}
\details{
The \code{Family} column is used to identify relevant columns.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
